requirements.txt
langchain
ollama
requests
python-dotenv
langchain_community
langchain-ollama
aiohttp